fix: planet curvature path
- Make sure all physics calculation is using double for better precision
- Fix stability issue with verlet
- Add exagerration factor for acceleration

+++ include/engines/maths/converter.hpp
@@ -5,6 +5,6 @@
 namespace Engines::Maths {
   class Converter {
     public:
-      static float meterToPixel(const float meter, const float *magnification = nullptr);
+      static float meterToPixel(const double meter, const double *magnification = nullptr);
   };
 }
