+++ include/engines/physics/physics_body.hpp
@@ -11,19 +11,19 @@ namespace Engines::Physics {
   template<typename T>
   class PhysicsBody {
     public:
-      T& setMass(float mass){
+      T& setMass(double mass){
         this->mass = mass;
         return static_cast<T&>(*this);
       }
-      T& setPosition(glm::vec3 position){
+      T& setPosition(glm::dvec3 position){
         this->position = position;
         return static_cast<T&>(*this);
       }
-      T& setVelocity(glm::vec3 velocity){
+      T& setVelocity(glm::dvec3 velocity){
         this->velocity = velocity;
         return static_cast<T&>(*this);
       }
-      T& setAcceleration(glm::vec3 acceleration){
+      T& setAcceleration(glm::dvec3 acceleration){
         this->acceleration = acceleration;
         return static_cast<T&>(*this);
       }
@@ -35,18 +35,18 @@ namespace Engines::Physics {
 
       virtual T& build() = 0;
 
-      float getMass() const {return mass;};
+      double getMass() const {return mass;};
       Engines::Graphics::Shape* getShape() const {return this->shape.get();}
-      glm::vec3 getAcceleration() const {return acceleration;};
-      glm::vec3 getVelocity() const {return velocity;};
-      glm::vec3 getPosition() const {return position;};
+      glm::dvec3 getAcceleration() const {return acceleration;};
+      glm::dvec3 getVelocity() const {return velocity;};
+      glm::dvec3 getPosition() const {return position;};
       
     protected:
-      float mass;
+      double mass;
       std::unique_ptr<Engines::Graphics::Shape> shape;
-      glm::vec3 position{0.0f, 0.0f, 0.0f};
-      glm::vec3 velocity{0.0f, 0.0f, 0.0f};
-      glm::vec3 acceleration{0.0f, 0.0f, 0.0f};
+      glm::dvec3 position{0.0, 0.0, 0.0};
+      glm::dvec3 velocity{0.0, 0.0, 0.0};
+      glm::dvec3 acceleration{0.0, 0.0, 0.0};
       std::vector<std::string> validation_messages;
 
       template<typename Func>
